Branch HEAD Use 64-bit XID field for async commit errors. b=2066 i=emoly.liu i=nathan.rutman

+++ lustre/include/obd.h
@@ -52,7 +52,7 @@ struct loi_oap_pages {
 struct osc_async_rc {
         int     ar_rc;
         int     ar_force_sync;
-        int     ar_min_xid;
+        __u64   ar_min_xid;
 };
 
 struct lov_oinfo {                 /* per-stripe data structure */
